requirements.txt
streamlit==1.37.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.3.2
joblib==1.3.2
matplotlib==3.8.4

shap==0.44.1

faiss-cpu==1.7.4

torch==2.1.2
torchvision==0.16.2

sentence-transformers==2.6.1
transformers==4.38.2
huggingface-hub==0.20.3


